Upgraded inkjs to v1.0!
That comes with the side effect of removing a previously unmentioned feature I'd patched into ink.js, namely code injection, which let you override functions in that file. I will add this back soon, but I don't think anyone was using it, and I'd rather upgrade to 1.0 as quickly as possible.